Changed cart to a Dictionary
