+++ WebShop/Controllers/DefaultController.cs
@@ -16,27 +16,6 @@ namespace WebShop.Controllers
             return View(Session["ListOfProducts"]);
         }
 
-        [HttpPost]
-        public ActionResult AddToCart()
-        {
-            var id = Request["addtocart"];
-            int index = int.Parse(id);
-
-            if(id != null)
-            {
-
-                List<Product> tmpList = (List<Product>)Session["ListOfProducts"];
-                tmpList[index].ReduceStockCount();
-                Session["ListOfProducts"] = tmpList;
-
-                Cart tmpCart = (Cart)Session["Cart"];
-                tmpCart.ProductList.Add(tmpList[index]);
-                Session["Cart"] = tmpCart;
-            }
-
-            return RedirectToAction("Index");
-        }
-
         [HttpGet]
         public ActionResult GetProduct(int id)
         {
