+++ WebShop/Views/Default/Index.cshtml
@@ -40,7 +40,7 @@
                         </form>
                     </td>
                     <td>
-                        <form action="/Default/AddToCart" method="post">
+                        <form action="/Cart/AddToCart" method="post">
                             <input type="hidden" name="addtocart" value=@i />
                             <button class="" type="submit">
                                 Add to cart
@@ -72,7 +72,7 @@
     <div class="row">
         <div class="container">
             <h2>Add product</h2>
-            <form class="form-group" action="~/Product/AddProduct" method="post">
+            <form class="form-group" action="/Product/AddProduct" method="post">
                 <input type="text" name="productName" placeholder="Name of product" />
                 <input type="text" name="productPrice" placeholder="Price" />
                 <input type="text" name="productStock" placeholder="Nr in stock" />
